pyproject.toml
[project]
name = "automated-auditor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "astor>=0.8.1",
    "astroid>=3.0.0",
    "black>=23.0.0",
    "docling>=1.0.0",
    "gitingest>=0.1.0",
    "gitpython>=3.1.40",
    "google-generativeai>=0.3.0",
    "isort>=5.12.0",
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-google-genai>=0.1.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.0.20",
    "langsmith>=0.1.0",
    "mypy>=1.5.0",
    "opencv-python>=4.8.0",
    "pdf2image>=1.16.0",
    "pillow>=10.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pymupdf>=1.23.0",
    "pypdf2>=3.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
    "ruff>=0.1.0",
    "typing-extensions>=4.8.0",
]
